chore(build): regenerate copilot-cli mirror for the claim-extraction split
Regenerates src/copilot-cli/skills/github/scripts/pr/ for the
closing_claim_extraction.py split (previous commit), and records the one
legitimate new scripts/repo/ reference in
.claude/skills/github/SKILL.md's Repository Settings section against the
skill_md_portability ratchet (issue #2050): baseline 10 -> 11 for both
.claude/skills/github/SKILL.md and its generated mirror. That file already
carries 10 identically-shaped scripts/<category>/ headers (PR Operations,
Issue Operations, Milestone Operations, and so on); this is the same
false-positive class the check's own UPSTREAM_PATTERNS comment names
(scripts/validation/check_skill_md_portability.py:256-261), one more
skill-relative scripts/ mention that never resolves to the upstream-only
top-level scripts/ tree the pattern is meant to catch.
