+++ scripts/validation/skill_md_portability_baseline.json
@@ -13,7 +13,7 @@
     ".claude/skills/cva-analysis/references/SKILL_SPEC.md": 1,
     ".claude/skills/doc-accuracy/SKILL.md": 1,
     ".claude/skills/git-advanced-workflows/SKILL.md": 1,
-    ".claude/skills/github/SKILL.md": 10,
+    ".claude/skills/github/SKILL.md": 11,
     ".claude/skills/memory-maintenance/SKILL.md": 2,
     ".claude/skills/observability/SKILL.md": 1,
     ".claude/skills/planner/SKILL.md": 3,
@@ -56,7 +56,7 @@
     "src/copilot-cli/skills/cva-analysis/references/SKILL_SPEC.md": 1,
     "src/copilot-cli/skills/doc-accuracy/SKILL.md": 1,
     "src/copilot-cli/skills/git-advanced-workflows/SKILL.md": 1,
-    "src/copilot-cli/skills/github/SKILL.md": 10,
+    "src/copilot-cli/skills/github/SKILL.md": 11,
     "src/copilot-cli/skills/memory-maintenance/SKILL.md": 2,
     "src/copilot-cli/skills/observability/SKILL.md": 1,
     "src/copilot-cli/skills/planner/SKILL.md": 3,
